@@ -2,19 +2,17 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-// @dart = 2.9
-
 import 'dart:async';
 
 import 'package:battery_platform_interface/battery_platform_interface.dart';
+import 'package:flutter_test/flutter_test.dart';
 import 'package:plugin_platform_interface/plugin_platform_interface.dart';
-import 'package:test/test.dart';
 import 'package:battery/battery.dart';
 import 'package:mockito/mockito.dart';
 
 void main() {
   group('battery', () {
-    Battery battery;
+    late Battery battery;
     MockBatteryPlatform fakePlatform;
     setUp(() async {
       fakePlatform = MockBatteryPlatform();
